refactor(components): [calendar]
* Extract date-table logic code into a dedicated file.
