+++ packages/components/calendar/index.ts
@@ -5,3 +5,4 @@ export const ElCalendar = withInstall(Calendar)
 export default ElCalendar
 
 export * from './src/calendar'
+export type { DateTableInstance } from './src/instance'
