+++ packages/components/calendar/src/date-table.ts
@@ -2,7 +2,6 @@ import { buildProps, definePropType, isObject } from '@element-plus/utils'
 import { rangeArr } from '@element-plus/components/time-picker'
 import type { ExtractPropTypes } from 'vue'
 import type { Dayjs } from 'dayjs'
-import type DateTable from './date-table.vue'
 
 export type CalendarDateCellType = 'next' | 'prev' | 'current'
 export type CalendarDateCell = {
@@ -47,5 +46,3 @@ export const dateTableEmits = {
   pick: (value: Dayjs) => isObject(value),
 }
 export type DateTableEmits = typeof dateTableEmits
-
-export type DateTableInstance = InstanceType<typeof DateTable>
