+++ packages/components/calendar/src/instance.ts
@@ -1,0 +1,3 @@
+import type DateTable from './date-table.vue'
+
+export type DateTableInstance = InstanceType<typeof DateTable>
